# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.DefaultResponse:
    properties:
      message:
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.SwagVacancyCategories:
    properties:
      name:
        type: string
    type: object
...
  models.VacancyCategory:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: This is a Tajik Career Hub API documentation.
  title: Tajik Career Hub API ✨
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/models.VacancyCategory'
            type: array
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.SwagVacancyCategories'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Retrieve a specific category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.VacancyCategory'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category with the provided details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.VacancyCategory'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/controllers.DefaultResponse'
        "400":
          description: Invalid ID or input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update category
      tags:
      - Categories
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"